pr: move test to preset_vllm
